Bug 716638 - Annotate the size of a failed allocation due to OOM and submit it with the crash report, r=ted
UltraBlame original commit: 4d6cb73f892bc8b8e10551e5b110f53cceadf1ba

+++ toolkit/crashreporter/google-breakpad/src/common/linux/linux_libc_support.h
@@ -105,7 +105,7 @@ my_strtoui(int* result, const char* s) {
 
 
 static inline unsigned
-my_int_len(int i) {
+my_int_len(intmax_t i) {
   if (!i)
     return 1;
 
@@ -125,7 +125,7 @@ my_int_len(int i) {
 
 
 static inline void
-my_itos(char* output, int i, unsigned i_len) {
+my_itos(char* output, intmax_t i, unsigned i_len) {
   for (unsigned index = i_len; index; --index, i /= 10)
     output[index - 1] = '0' + (i % 10);
 }
